code/webapp/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0
langchain==1.2.10
langchain-openai==0.2.11
langchain-community==0.3.14
langchain-text-splitters==0.3.4
qdrant-client==1.7.0
pypdf2==3.0.1
youtube-transcript-api==0.6.2
